target: add multi DTB image generation support
This changes allows to generate a single itb image
with multi dtb configuration for ipq95xx and ipq53xx

The change defines a new board type qcom_alxx,which
when selected as part of profile, will generate single
itb image with multi dtbs that are present in KDIR
for the particular soc.

The lzma compression support is included for this itb
image.

Change-Id: Id95bd2b3192ca6eaa115828396dfae98a782904a
Signed-off-by: Ermine Jose <quic_erminejo@quicinc.com>